fix(gui): cleanup remaining hardcoded colors in app chrome and file player
- Margin::same cast to i8 for egui API compat
- file_player.rs: 5 Color32::from_rgb replaced with SonidoTheme references

+++ crates/sonido-gui/src/app.rs
@@ -652,7 +652,7 @@ impl SonidoApp {
             .fill(theme.colors.void)
             .stroke(Stroke::new(1.0, theme.colors.amber))
             .corner_radius(theme.sizing.panel_border_radius)
-            .inner_margin(Margin::same(theme.sizing.panel_padding));
+            .inner_margin(Margin::same(theme.sizing.panel_padding as i8));
 
         let panel_response = panel_frame.show(ui, |ui| {
             ui.set_min_height(160.0);
